Temporary work-around: don't loose target while creating an annotation.
See #1020 for a description of the problem.

This commit patches the merge behavior of the bridge plugin:
never update a list of targets with a shorter list.
(This avoids it being zeroed by accident.)

This does not really solve the more generic problem: the race condition
of the messages moving on the bridge. We should either do some
time-stamping, or some other smarter flow control...